Fixing lazy transactions

+++ src/Voron/Impl/Journal/IJournalWriter.cs
@@ -7,7 +7,7 @@ namespace Voron.Impl.Journal
     public unsafe interface IJournalWriter : IDisposable
     {
         void WriteGather(long position, IntPtr[] pages);
-        long NumberOfAllocatedPages { get;  }
+        int NumberOfAllocatedPages { get;  }
         bool Disposed { get; }
         bool DeleteOnClose { get; set; }
         IVirtualPager CreatePager();
